[WIP] Fix 3D shadows not being drawn in viewports by default [ci skip]
This sets the default shadow atlas size in Viewport to 4096
(similar to the default value in the Project Settings).

This also makes Godot always call lower-level functions when using
`Viewport::set_shadow_atlas_size()`. This is a hack to make sure
the shadow atlas is created with the correct resolution, as merely
changing the default setting didn't fix the issue.

Edit: After some testing, this commit doesn't seem to fix things
anymore, I don't know why.

This closes #27650.