[WIP] Add a keyboard shortcut to toggle editor debug draw mode [ci skip]
This shortcut toggles between the Normal debug draw mode and the last
manually selected debug draw mode, or Wireframe if no mode was
manually selected.

This is useful to quickly compare different views with debug draw
modes, or seeing through walls temporarily.

TODO:

- Fix inconsistency when storing state. I don't know why it's happening.
  Pressing the toggle key multiple times after starting the editor results in this:

Going to 20
Storing 23
Going to 23
Storing 21
Going to 21
Storing 22
Going to 22
Storing 20
Going to 20
Storing 23
Going to 23
Storing 21
Going to 21