[WIP] Use fake directional lighting for fully baked surfaces [ci skip]
This makes normal maps useful on fully baked surfaces that don't get lit
by real-time lights.

This looks good if you only use a DirectionalLight, but will look bad
on fully baked OmniLights/SpotLights since they will barely light up walls
depending on their orientation.

This is best tested with a DirectionalLight that has its bake mode set to All
and rotation set to `(-0.5, 0.5, 0)` (in radians).

The code only modifies what's needed for GLES3 (and for atlased lightmaps),
but it could be adapted to non-atlased lightmaps and/or GLES2.