[WIP] Improve normal and AO blending logic in the decal shader [ci skip]
This makes normal and AO blending behave more logically.
Instead of replacing the underlying surface's normal map entirely,
the decal will now add to the existing normal map and progressively
replace it as the Albedo Mix value is increased.

For ambient occlusion, the lowest value between the underlying surface's
AO map and the decal's AO map will be chosen (in a way similar to AO/SSAO
interactions).

TODO:

- Use UDN normal map blending for correct results.
- Fix AO blending.