[WIP] Add immutable/one-time assignable variables to GDScript (`let`) [ci skip]
Both local variables and member variables can be made immutable
by using the `let` keyword instead of `var` during the declaration.

This can be used to write safer code in situations where `const`
is not suited (as `const` is limited to constant expressions).

This `let` keyword behaves in a way similar to Nim's `let` keyword,
Kotlin's `val` keyword and JavaScript's `const` keyword. The immutability
is "flat" in the sense that the underlying data can still be modified
(e.g. arrays and dictionaries can be modified, but not reassigned).

Immutable variables must also be initialized as soon as they're declared.

TODO:

- Fix member variables that can still be reassigned even if immutable.
- Fix incorrect line number for error if an immutable variable isn't
  initialized immediately after it's declared.
- Disallow setters and getters for immutable variables.