Tweak point light shadow near plane [ci skip]
This improves the effective depth buffer precision for
point light shadows. As a side effect, lights with a very large
range will no longer cast shadows for very close objects. This can
be worked around by moving the light or the object.

This matches the current `master` implementation of the near clip for
point light shadows.