[WIP] Add a step resolution property for screen-space reflections
This can be lowered below the default value (0.5) to improve
performance, or increased above the default value to improve quality.

The number of steps is automatically adjusted behind the scenes,
so that the length of screen-space reflections remains the same
(as it is generally the result of an artistic decision).

TODO:

- Fix SSR effect not doing anything.