[WIP] Add Y offset to screen-space reflections
This helps make background lines on floor/wall intersections
less noticeable, at the cost of making such lines more noticeable on
ceiling/wall intersections. Since most of the player attention
(and lighting) generally goes towards the floor and lower walls,
this is usually a good tradeoff. It may be possible to use a
normal-based approach to avoid this caveat.

TODO:

- Add an Environment setting to control the Y offset (range: 0,1,0.01).