Fix ReflectionProbe intensity blending with other GI sources
When ReflectionProbe intensity is below 1.0, the reflection and ambient
light's intensity are now properly blended with other GI sources
(such as VoxelGI, SDFGI or environment sky). Previously, reflections
would be darkened and ambient lighting was never reduced.

This allows for smooth fading of ReflectionProbes over distance,
which is useful for LOD purposes.

FIXME:

- Prevent overlapping reflection probes from rendering too bright.
  Maybe a separate `intensity_accum` inout parameter is needed for
  `reflection_process()`?