[WIP] Add caret width editor setting and LineEdit/TextEdit caret width settings
This can be used to improve caret visibility in LineEdits and TextEdits
used by the editor (including the script editor).

In the editor, the default caret width is now identical for both LineEdit
and TextEdit (instead of TextEdit having a 2 pixel-thick caret).
Projects are not affected by this default setting change.

TODO:

- Make it work for TextEdit (it only works for LineEdit right now).