[WIP] Implement distance fade in OmniLight and SpotLight
This can be used to improve performance in large scenes by fading
distant lights or their shadows. This works in both GLES3 and GLES2.

Unlike the `master` implementation, this supports smoothly fading
the shadow in the distance since colored shadows are supported.
The shadow fade distance is equal to the light's fade distance.

TODO:

- Cull light rendering when out of range in GLES3
  (it doesn't seem to be done correctly?).
  - Shadow rendering cull seems to work though.
- Cull light and shadow rendering when out of range in GLES2.