[WIP] Use high-quality graphics settings by default when Movie Maker is enabled
This results in much greater visual quality out of the box when
recording videos with Movie Maker functionality.

Most settings are maxed out, with the exception of shadow size
which is kept to 8192 to keep a similar shadow sharpness between
real-time and movie rendering (as Ultra soft shadow quality doubles
the blur radius for constant-blur shadows).
This also avoids using too much VRAM and slowing down rendering as a result.

Settings that have improved filter quality such as SSAO and SSIL
have their resolution and blur passes increased for a similar
visual appearance (but with greater stability).

Supersampling is also not used as it requires too much VRAM for most
GPUs at high recording resolutions.

TODO:

- Settings don't seem to apply when Movie Maker is enabled.