[WIP] Add a Clearcoat Normal Texture property to BaseMaterial3D
This allows overriding the direction of the clearcoat effect, similar
to a normal map for standard specularity.

For best results, a dedicated normal map should be authored to be used
with clearcoat. Clearcoat is originally intended to purely follow geometric
detail, but when using lowpoly meshes, this detail may not be sufficient.

TODO:

- Make it look correct.
- Update shader when clearcoat normal texture is set.
- Port to Forward Mobile and Compatibility rendering methods.