[WIP] Add virtual method to run code when a new object is instantiated in the editor
This is editor-only by design and does not affect objects already present in
a scene when loading it.

This can be used to automatically set up nodes/resources in the editor
to reduce the number of clicks required to get to a workable result,
without breaking compatibility with existing projects.

TODO:

- Make it work (the receiver method is currently not actually called).