[WIP] Make Run Project and Run Current/Specific Scene shortcuts effective on a running project
Like Stop Running Project (F8), these shortcuts are now effective
when the project window is focused. Since the project is always running
by definition in this situation, the project will be restarted on the main
scene (F5), current scene (F6) or the specific scene that was previously
started (Ctrl + Shift + F5).

This also refactors the debugger shortcut code to avoid repetition
and bind macOS shortcuts as fallbacks.

TODO:

- Figure out why the project is not restarting when pressing F5 or F6.
- Fix save conflict dialog appearing when pressing F5 or F6
  while project is running. This does not occur when Save Before Running
  is disabled in the Editor Settings.
- Make Run Specific Scene functional by reloading the specific scene
  that was previously run. If no specific scene was previously being
  run, show the Run Specific Scene dialog.