Import resources when running a project that tries to load missing resources
This works by restarting the engine with a special command line argument
that works like `--import`, but runs the project immediately after
the editor is done importing resources (and closes afterwards).

This makes it much more seamless to run projects from the command line,
as you no longer have to manually ensure `.godot/` is fully imported
and up-to-date before you run the project.