[WIP] Add a capability system disallow using certain features from scripts
This can be used to harden a Godot project against untrusted code,
reducing its potential to damage the system it's running on or compromise
the user's privacy.

Capabilities are intended as a first line of defense, and are not
intended to replace full-blown sandboxing solutions or other
security practices.

All capabilities are enabled by default to preserve existing behavior.
Four capabilities are currently available to disable:

- Filesystem access (with a list of allowed paths when disabled).
- Network access (with a list of allowed addresses when disabled).
- Clipboard access.
- Process management.

This system is called "capabilities" to avoid confusion with OS permissions,
as used on the Android, iOS and web platforms.

TODO:

- Implement filesystem, network and clipboard restrictions.
- Determine whether `OS.shell_open()` and `OS.shell_show_in_file_manager()`
  should be still blocked when Allow Process Management is `false`.
- Prevent project settings in the `application/capabilities/` category
  from being overridden using `override.cfg`.
- Prevent `ProjectSettings.save()` from overriding existing settings
  in the `application/capabilities` category (for `@tool` scripts and editor plugins).