[WIP] Add a Shadow Jitter property to Light3D
This jitters the shadow rotation every frame to improve quality
when TAA or FSR2 is enabled. This is particularly helpful to avoid visible
stairstepping in the shadow map, regardless of the shadow filter quality chosen.
The effect is most noticeable with dynamic objects that move slowly.

The downside is that the shadow map must also be updated every frame,
even if nothing has actually changed within its bounds. This can have a negative
performance impact if the light's shadow map could have been cached otherwise.
For DirectionalLight3D (which usually has to be updated every frame
due to containing dynamic objects), this shouldn't make a performance difference
in most scenes still.

Unlike GDScript-based solution where the Light3D node is rotated,
this doesn't affect the projector texture and doesn't affect child nodes
of the Light3D.

This is only implemented in Forward+, as it requires TAA or FSR2 to have a positive
effect on shadow quality.

TODO:

- Make it effective only in Forward+ (as Mobile lacks TAA/FSR2), and only
  when TAA or FSR2 is enabled in the viewport.
- Fix OmniLight3D/SpotLight3D flickering all over the place when Shadow Jitter
  is enabled.